Add prev_root to node
I was reading ahead in the Intermediate Git lessons and the explanation
of how commits and pointers work reminded me of a linked list, but
backwards. Commits point to their parent instead of vice versa.
I think that could work here with tracing back the steps it took to get
to the end position. This isn't a binary tree, so I can't start from the
root and go down very easily.